requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.21
python-dotenv==1.0.0
gunicorn==21.2.0